xc refactor : move relevant files to module_xc
