+++ source/CMakeLists.txt
@@ -5,6 +5,7 @@ add_subdirectory(module_neighbor)
 add_subdirectory(module_orbital)
 add_subdirectory(module_md)
 add_subdirectory(module_deepks)
+add_subdirectory(module_xc)
 add_subdirectory(src_io)
 add_subdirectory(src_ions)
 add_subdirectory(src_lcao)
