+++ source/module_xc/xc_functional.cpp
@@ -1,7 +1,6 @@
 #include "xc_functional.h"
 #include "xc_type.h"
-#include "global.h"
-#include "myfunc.h"
+#include "../src_pw/global.h"
 #include "../module_base/global_function.h"
 #include <stdexcept>
