+++ source/module_xc/xc_type.cpp
@@ -1,6 +1,6 @@
 #include "xc_type.h"
 #include "../module_base/global_function.h"
-#include "global.h"
+#include "../src_pw/global.h"
 #include "exx_global.h"
 
 xcfunc::xcfunc()
